Move model overview components into their own files
